backend/requirements.txt
# Database
psycopg[binary]>=3.1
python-dotenv==1.0.0


# Authentication
passlib==1.7.4
bcrypt==4.0.1
Flask-JWT-Extended==4.6.0

# Data validation
pydantic>=2.6.0

# Utilities
pytest==7.4.0
python-dateutil==2.8.2
gunicorn==21.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Flask
Flask==2.3.2
Werkzeug<3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
flasgger==0.9.7.1
PyJWT==2.8.0 


